requirements.txt
crewai>=0.28.8
crewai[tools]
openai>=1.0.0
anthropic>=0.28.0
groq>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0
moviepy>=1.0.3
elevenlabs>=1.0.0
youtube-data-api>=0.0.20
pillow>=10.0.0
pydantic>=2.0.0
